Adding info/error bar functionality
Can flash message as either error and info and have it render as
and alert box.  Clicking the box will close it.  Maybe add a UI
element prompting the user that it can be closed.  Maybe also add
a timer to have the box fade out and close after X time?